refactor(telegram): split actions.ts into action modules
actions.ts was 759 lines — one createTelegramActionHandler factory
wrapping a ~660-line switch over 40+ Bot API actions. Split into an
actions/ folder using the same registry pattern as gateway-actions:
  - types     — TelegramActionContext + handler-map types
  - shared    — replyParams + sendText
  - messaging — send/reply/edit/delete/pin/forward/react/schedule/buttons
  - media     — files/photos/videos/stickers/polls/locations/contacts/dice
  - chat-info — chat metadata, userbot history, sticker packs, downloads

index.ts merges the maps and rebuilds createTelegramActionHandler with
the identical signature; the per-instance scheduledMessages timer state
moves onto a shared context object. sendText still exported.
Verified: tool-functional suite passes (44 tests).
